moved files and renamed booleans
